better loading of package code when releasing codec dpkgs
the added boilerplate code at the top of each script that makes a
codec data package checks to see if R is running in a directory with
the CoDEC DESCRIPTION file.  If it is, we assume that we want to load
the development version of codec instead of the `library` version of
codec.

this will allow anyone to still use these scripts to regenerate data
without having to install devtools and clone the repo